enable `EBITENGINE_INTERNAL_IMAGES_KEY` on browsers Fixes #2270
