+++ internal/restorable/images_js.go
@@ -1,4 +1,4 @@
-// Copyright 2022 The Ebitengine Authors
+// Copyright 2017 The Ebitengine Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,32 +12,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package graphicscommand
+package restorable
 
 import (
+	"archive/zip"
 	"bytes"
 	"image"
-	"path/filepath"
 	"syscall/js"
 
 	"github.com/hajimehoshi/ebiten/v2/internal/graphicsdriver"
 )
 
-func (i *Image) Dump(graphicsDriver graphicsdriver.Graphics, path string, blackbg bool, rect image.Rectangle) error {
-	// Screen image cannot be dumped.
-	if i.screen {
-		return nil
+// DumpImages dumps all the current images to the specified directory.
+//
+// This is for testing usage.
+func DumpImages(graphicsDriver graphicsdriver.Graphics, dir string) error {
+	buf := &bytes.Buffer{}
+	zw := zip.NewWriter(buf)
+
+	for img := range theImages.images {
+		f, err := zw.Create(img.image.DumpName("*.png"))
+		if err != nil {
+			return err
+		}
+
+		if err := img.image.DumpTo(f, graphicsDriver, false, image.Rect(0, 0, img.width, img.height)); err != nil {
+			return err
+		}
 	}
 
-	path = i.DumpName(path)
-	path = filepath.Base(path)
+	zw.Close()
 
 	global := js.Global()
-
-	buf := &bytes.Buffer{}
-	if err := i.DumpTo(buf, graphicsDriver, blackbg, rect); err != nil {
-		return err
-	}
 
 	jsData := global.Get("Uint8Array").New(buf.Len())
 	js.CopyBytesToJS(jsData, buf.Bytes())
@@ -45,10 +51,10 @@ func (i *Image) Dump(graphicsDriver graphicsdriver.Graphics, path string, blackb
 	a := global.Get("document").Call("createElement", "a")
 	blob := global.Get("Blob").New(
 		[]interface{}{jsData},
-		map[string]interface{}{"type": "image/png"},
+		map[string]interface{}{"type": "archive/zip"},
 	)
 	a.Set("href", global.Get("URL").Call("createObjectURL", blob))
-	a.Set("download", path)
+	a.Set("download", dir+".zip")
 	a.Call("click")
 
 	return nil
